requirements.txt
ipython~=8.6.0
tensorflow~=2.9.0
matplotlib~=3.5.3
numpy~=1.21.5
keras~=2.9.0
scikit-learn~=1.1.3
pillow~=9.2.0